feat: add bin config modal
